+++ moshi/moshi/exceptions.py
@@ -1,0 +1,3 @@
+class SetupError(Exception):
+    """Raised when there's a problem with the state monad."""
+    ...